+++ packages/node/src/env.ts
@@ -2,11 +2,7 @@ import type { Canvas as NodeCanvas } from 'canvas';
 import { JSDOM } from 'jsdom';
 // @ts-expect-error internal import
 import utils from 'jsdom/lib/jsdom/living/generated/utils.js';
-import type {
-  TCopyPasteData,
-  TFabricEnv,
-  TFabricWindow,
-} from '../../../src/env/types';
+import type { TCopyPasteData, TFabricEnv, TFabricWindow } from '@fabricjs/core';
 import { NodeGLProbe } from './NodeGLProbe';
 
 const { implForWrapper: jsdomImplForWrapper } = utils;