+++ vitest.setup.ts
@@ -1,5 +1,5 @@
 import './vitest.extend';
-import { setEnv, setEnvFactory } from './src/env';
+import { setEnv, setEnvFactory } from './packages/core/src/env';
 import { getEnv } from './packages/browser/src/env';
 import { beforeAll } from 'vitest';
 import { isJSDOM } from './vitest.extend';